WL#12805 :  ClusterSet Switchover and Failover
This patch introduces the following maintenance operations of
ClusterSets:

  - Failover of Primary Cluster
  - Planned switch of the Primary Cluster
  - Rejoin of a Cluster

The API was extended with:

  - <ClusterSet>.forcePrimaryCluster(), to perform a failover of the
Primary Cluster of the ClusterSet, ensuring the target Cluster is
promoted to Primary and the previous Primary Cluster is invalidated as
well as several consistency checks and proper update of the replication
channel.

  - <ClusterSet>.setPrimaryCluster(), to perform a safe switchover of
the Primary Cluster of a ClusterSet. It ensures the current PRIMARY is
demoted to a Replica Cluster, while the promoted Cluster will be made the
Primary Cluster. All other Replica Clusters are updated to replicate from
the new Primary. The function also ensures the GTID-sets of the Clusters
in the ClusterSet are reconciliated when purged transactions as being
VCLEs (view change log events) are detected.

  - <ClusterSet>.rejoinCluster(), to rejoin back to a ClusterSet a
Cluster that was previously invalidated during a failover or switchover.
It also can be used to restore/update the replication channel in Replica
Clusters if it does not have the expected state. The function ensures
the Cluster complies with the requirements to be rejoined back to the
Cluster before performing the operation.

On top of that, this patch also enables CLI integration of ClusterSets

Change-Id: I7fdaa1ac434e5f9bbe3f8a4404028e128b52716f