BUG#32226871 : MISSING INFORMATION ABOUT METADATA NOT INCLUDING SERVER_ID ON UPGRADE SCENARIO
In 8.0.23, a check was added to verify if the server_id of all cluster
members is registered in the metadata as an instance attribute and if
not the metadata is updated accordingly.

This check is executed on add, rejoin and rescan operations.

However, when upgrading a cluster from a version previous < 8.0.23
to a version >= 8.0.23, the server_id won't be registered in the
metadata unless the user performs a manual rejoin. This is silently
ignored as it's not included in any diagnostics.

This patch fixes that by adding a new verification to check if the
server_id of cluster members is missing from the metadata and includes
a NOTE message in the 'instanceErrors' attribute of the output of
.status() indicating to use .rescan() to fix it.

Added regression tests to verify the changes.

Change-Id: I0dd7a7450c98b2d6265e224f0de930c6a3b79bf3