BUG#32461414 BUG#32157120 BUG#29779995 BUG#30618067 Adding CLI Support for Cluster/Replicaset
BUG#32461414 CLI HELP FOR CLUSTER/REPLICASET FUNCTIONS IS MISSING OPTION
DETAIL

This patch replaces the Dictionary options in the Cluster functions for
Option Packs. The result of this is full CLI support for functions
in the Cluster object, including the CLI help integration.

BUG#32157120 ADMINAPI BUILT-IN HELPER UNNECESSARILY DESCRIBES COMMANDS
EXCEPTIONS

This patch removes the exceptions from the API functions in Clusters and
Replicasets.

BUG#29779995 ADDINSTANCE() ALLOWS THE OPTIONS EXPELTIMEOUT AND CONSISTENCY
WHEN IT SHOULDN'T
The expelTimeOut and consistency options were removed from Cluster.addInstance

BUG#30618067 MYSQLSH -- CLUSTER HELP DOES NOT PROVIDE HELP
This patch is exactly about solving this problem which required, however the
syntax is slightly different:

    $ mysqlsh -- cluster --help

The expelTimeOut and consistency options were removed from the
Cluster.addInstance API.

Change-Id: I0d8784073e2e411c2c621ecfcfa465a2b13db8a9