BUG#31961688 RESUMING LOAD OF SUB-CHUNKED DUMPS DOESN'T SKIP ALREADY LOADED ROWS
When resuming load of a dump which contained chunks that are too big and
needed to be sub-chunked, the rows which were already loaded were not
skipped, which might caused deadlocks.

All sub-chunks are now recorded in the progress file, if the load
operation is interrupted and resumed, sub-chunks which were loaded
successfully are skipped.

Change-Id: I4391c0bf6910bc471c72b83d6584d5aa9ff680b8