BUG#29037274 : ADMINAPI MUST NOT SET DEFAULT VALUE FOR G_R_EXIT_STATE_ACTION IN 8.0.15
With the patch for BUG#28701263, the AdminAPI started setting and
persisting a default value for group_replication_exit_state_action
being that 'READ_ONLY'. The default used by group replication was
'ABORT_SERVER'.

However, in MySQL Server 8.0.16 the default value of
group_replication_exit_state_action became 'READ_ONLY' so the AdminAPI
should not change and persist it.

This patch fixes that by ensuring the setting is not changed when the
target instance version is >= 8.0.16.

Added regression tests.

Change-Id: Ibc47ce8d9693037ce91043bd044b31aa93d0c9d0