BUG#31991122 SHELL ADD EMPTY FILE INFORMATION USING UTIL.IMPORTTABLE WITH OCI
When object is created explicitly as directory in OCI Object Storage
bucket, then it is later listed as file entry with 0 byte size and '/'
postfix.

If we list all objects from such directory using wildcard, e.g. logs/*,
then logs/ directory entry is also included in object list.

Now we are checking if object entry is a directory and we filter out
such entries when doing wildcard matching on files.

Change-Id: I91162a047a46ccdcad0761632b43d4a36b60ccd1