BUG#32157182 RESCAN() DOESN'T VERIFY IF THE INSTANCES RPL USERS ARE MISSING FROM THE METADATA
With this patch we try to detect problems related to recovery accounts.
The status() call will print a message if the recovery account is missing
in the MD schema as also if it's not the one actually used. User will be
instructed to either re-add the instance or run rescan() based on the actual
detected situation.
The addInstance() will also print a hint to call rescan()
if some missing recovery users will be found in the MD schema.

Change-Id: I1906df4d605747c8bda7a6681f7c67659a6eecb1