BUG#32220514 SHELL.LIST (PYTHON ARRAY WRAPPER) IS NOT COMPATIBLE WITH PYTHON STDLIB LIST()
The implementation of the array wrapper used by Shell in Python mode has
been adjusted, it now inherits from the Python's list (isinstance()
checks will return true). In order to avoid duplication of data and
ensure synchronization between Python's context and internal
representation of data, data is not stored in the base class, but in the
array wrapper. Due to this fact (and the fact that data is shared
between Python and JavaScript mode) it is only safe to store basic types
(i.e. numbers, strings, arrays, dictionaries) and native Shell objects
(i.e. Cluster, Session). As the base class is not used to store the
data, all methods of the list class have been reimplemented.

Change-Id: Ib0f3f26eed9676d0c1ea39108e3262588adaa0e8