BUG#32241000 BUG#32247631 : ADD SUPPORT FOR VERIFY_CA AND VERIFY_IDENTITY SSLMODE
VERIFY_CA and VERIFY_IDENTITY were not supported in memberSslMode for
the following commands:

- dba.createCluster()
- <Cluster>.addInstance()
- <Cluster>.rejoinInstance()

This patch fixes that by introducing support for those options as well
as a validation to ensure that whenever those are used, the CA
certificates are supplied via ssl-ca or ssl-capath.

Updated and added new unit-tests to verify the changes.

Based on a code contribution from Daniël van Eeden, tracked by:
Bug#101914

Change-Id: I0e86ba933eb59c4cff82afd6c24eea77614d569a