BUG#32287986 CHECKINSTANCECONFIGURATION() PROVIDES INCORRECT INFORMATION ABOUT MISSING GRANTS
dba.checkInstanceConfiguration() was performing a wrong validation
regarding the required privileges with grant option. This resulted in an
endless loop on which the user would give the grants as informed by the
command and the command would keep failing with the indication the
grants were missing.

The root cause of this problem was the internal verification of the
privileges with grant option which was using the full list of available
grants instead of the list of grants required for the InnoDB
Cluster/ReplicaSet administration user. This patch fixes that by
ensuring the verification is done using the right list.

On top of that, the internal list of required grants included SUPER
which is deprecated in 8.0. This patch fixes that by replacing the SUPER
grant with the fine-grained grants.

Added new tests to verify the fix.

Change-Id: I40059f8f1b6861394cdd43fa1cd89f5db642361d