BUG#32515696 SHELL DUMP*(): SUPPORT NONGTID ENABLED/SUPPORTED
If dump was executed on a server which does not have GTID_EXECUTED
system variable, or does not have information_schema.COLUMN_STATISTICS
table, it would fail, even though this information are not essential
for a successful dump.

This is now fixed, in case of problems described above, a warning
message is going to be displayed and a detailed error message is going
to be logged.

Change-Id: Iec533484c63a189f623370b2f3c076ff7b7edbf3