BUG#32528186 BUG#32540460 Refactoring of Instance_cache
BUG32528186 MYSQLSH UTIL.DUMPINSTANCE CAUSES THREAD STACK OVERRUN ON SERVER

If MySQL server instance contained a large number of objects (i.e.
schemas, accounts, etc.), dump failed with thread stack overrun errors.

Queries which gather information about instance aim to fetch as much
information as possible using the least number of queries. This
behaviour has been modified, in case of errors such as reported in this
bug, query is going to be repeated, fetching information for a smaller
number of objects.

Additionally, all queries have been rewritten in order to avoid ORDER BY
clauses which could trigger filesort, allowing to run dump operation
even in case of a server running on a host with a full disk.

BUG32540460 DUMP FAILS IF AN EMPTY SCHEMA IS INCLUDED

When using dump utilities, if instance could not be locked with FLUSH
TABLES WITH READ LOCK and dumper had to fall back to LOCK TABLES, if
instance contained a schema which did not have tables, dump operation
might have failed.

If schema which had no tables happened to be the last on the list of
schemas to be dumped, internal flag was set to a wrong value and dump
assumed that there are no tables to be dumped. The table metadata was
not fetched, resulting in syntax errors in queries which fetched table
data.

The internal flag causing problems has been removed as part of the fix
for BUG#32528186, added regression test.

Change-Id: I96f77ea08d9436dbb4e71b3aae816d30e80c135f