BUG#32582745 ADMINAPI: OPERATIONS LOGGING USELESS MD STATE INFORMATION
Most of the Admin API operations contain Metadata preconditions to determine
if it is valid or not to execute them. When such check is done, it was adding
2 entrires on the log, one to indicate the check was about to be done and
another to indicate the metadata state. The found MD state is basically useless
unless it was not OK. Having these entries logged all the time causes the log
to get polluted, specially when monitoring a cluster which implies executing
status request in a recurring way.

This patch merges logging to of the mentioned entries in a single entry that
gets logged as info when the state is not OK and as debug info when the state
is OK (so only logged is log_level==debug).

Change-Id: Ia34a388e713a9af47656e38d03fa94de13a51fc8