BUG#32587098 CRASH WHEN EXECUTING TESTS
When loading a dump, there was a possibility of a crash if task data was
accessed after task itself was deleted.

Reordered the code to make sure function calls which may cause tasks to
be deleted to be executed last, reworked a function which can be called
after task is deleted so that it does not use task data.

Change-Id: I37c5ac99a1552008521661cb6feb260ffece3cdb