BUG#32528110 UTIL.DUMPINSTANCE CAUSES SEGFAULT IN MYSQLSH
If an exception is thrown from the main dumper thread while the worker
threads are starting, and one of them is slow to start, Shell would
crash.

When an exception was thrown from the main thread, the session in that
thread was closed and deleted. If one of the worker threads was slow to
start, it would access the session of the main thread which was already
deleted and dereference a null pointer. In order to prevent this issue,
main session is no longer deleted after it is closed, memory is freed
when the whole dumper class is deleted.

Change-Id: I9ef607de6917c8fc68045ffde7847b49649383d0