Fix random failures in forceQuorum() tests
The current tests for forceQuorumUsingPartitionOf() do not take into
account instances with GR settings persisted that automatically start
GR and try to join the group.

In a very fast machine, this can trigger a race condition that causes
the failure of the tests. The command tries to obtain the list of
instances belonging to the partition of the target instance to use it
to set group_replication_force_members. However, in such scenarios that
list may contain an instance with the status RECOVERING that will
be immediately kicked-out but is used nevertheless.

This patch fixes that by ensuring GR's startup procedure has completely
finished before running forceQuorum()

Change-Id: I591de8be61120b4f9a6457e6123e97a10b7abbaf