WL#14506 Dump/Load compatibility options for handling missing PKs
Changes in behavior
===================

When a dump is created and the `ocimds` option set to true, for each
table that would be dumped which does not contain a primary key, an
error is reported.

New values supported by the `compatibility` dump option
=======================================================

ignore_missing_pks
------------------

Suppresses the errors described above, MDS HA cannot be used with this
dump.

create_invisible_pks
--------------------

Each table without a primary key will have it created automatically when
the dump is loaded, Inbound Replication into an MDS HA instance cannot
be used with this dump (at the time of the release of MySQL Shell 8.0.24).

New option supported by the `util.loadDump()`
=============================================

createInvisiblePKs
------------------

Boolean, the default value depends on the dump being loaded: if it was
created using the `create_invisible_pks` value given to the
`compatibility` options it's `true`, otherwise it's `false`.

If set to `true`, primary keys are going to be automatically created
when the table DDL is executed: either using the functionality
described in WL#13784 (in case of a server which supports it), or by
modifying the `CREATE TABLE` statement and adding a
`my_row_id BIGINT UNSIGNED AUTO_INCREMENT INVISIBLE PRIMARY KEY`
column to the table.

If server has version lower than 8.0.24, load command will fail.

Change-Id: I8490fd73fcf1e523c5f92a50e4fbf60a647b062c