BUG#32603101 FIXING TESTS AFFECTED BY CHANGE IN HANDSHAKE IN CLASSIC SESSIONS
The logic in libmysql to determine the authentication plugin to be used in
the hello packet has changed.

Prior to 8.0.25 the possible scenarios include:

a) No auth-method was specified: in such case, the
caching_sha2_password_client_plugin was used.
b) An auth-method was specified and it was recognized, so it was used
c) An auth-method was specified and but was not recognized: the connection
failed with: "Authentication plugin '<name>' cannot be loaded"

In 8.0.25 the server sends the client-side authentication plugin name as part
of the hello packet, so the logic to pick the default authentication method
to be used changed to:

d) If the plugin name sent by the server is recognized, it is used
e) If the plugin name sent by server is not recognized, then follow the logic
as before 8.0.25

This change are causing some tests that before were on scenario c) are now
falling on scenario d)

This patch fixes the expectation on the failing tests.

Change-Id: I55d8755cc657ddaf632cba575f2d4fc4b61144c0