Fix random failures of dba_cluster_force_quorum test:
In 5.7 the test would fail randomly due to BUG#26394418 (Group
Replication) that was fixed in 8.0 but not backported to 5.7.

The bug happens more frequently when the loss of quorum is triggered by
killing the remaining cluster members and restarting them. If at the
moment the command to restore quorum is executed it's possible to
connect to those members, the command can fail with a timeout.

To workaround such failure, we simply don't restart those instances
after making the cluster loose the quorum.

This does not affect the test purpose and outcomes.

Change-Id: Id7c8d00719122a4ce73f43e93fc36bde6de97b62