WL#13815 : AdminAPI - Cluster object updates for ClusterSet
This patch adds support for InnoDB Cluster operations in Clusters that
are part of ClusterSets

It ensures that:

  - Metadata operations are performed in the right Cluster (Primary)
  - The replication channel between Clusters is kept healthy and
functional when operations need to change it or restore it
  - Replica Clusters are not writable by clients
  - Cluster commands have the correct availability depending on the
ClusterSet and individual Clusters status, i.e. operations may be
unavailable or not

On top of that, it also:

  - Ensures .getCluster([clusterName]) returns a handle to the Cluster
named <clusterName> part of the ClusterSet the current Shell session is
established to
  - dba.createCluster() generates and sets a value for
'group_replication_view_change_uuid' to be used by the Cluster
  - <Cluster>.rescan() detects if 'group_replication_view_change_uuid'
is enabled or not in the target Cluster and updates it accordingly to
ensure the Cluster is ready for ClusterSet
  - <Cluster>.addInstance()/.rejoinInstance() enables
'skip_replica_start' and uses the value of
'group_replication_view_change_uuid' set at Cluster creation time and
configures the ClusterSet replication channel accordingly.
  - <Cluster>.removeInstance() resets all members actions and the
ClusterSet replication channel settings

Change-Id: Ie5744286a064be6c4144b9e6d7733b57974e5866