BUG#32470621 SHELL JS CONSOLE MAY LEAVE VARIABLE IN DOOMED STATE FOREVER
In JS mode, if variable is defined using the `let` keyword, but is not
properly initialized due to i.e. exception, it is left in a state where
it cannot be accessed or properly initialized.

This patch changes the way the interactive JS statements are executed,
the V8's REPL mode is now used. This mode allows for redefinition of
`let` variables, meaning the problem described in the bug can be fixed
by simply defining the same variable with a different value. Any
variable can be redefined, not only the uninitialized ones.

The REPL mode currently does not allow to redefine the `const`
variables, hence in case of these variables issue remains unfixed.

Change-Id: I42bb4181a39c9cf2c9c2ed02334e0054bb4804f2