BUG#32741098 BUG#32788788 Fixes for the dump utilities
BUG#32741098 SHELL DUMP: USERS WITHOUT PASSWORD SHOULD BE FILTERED OUT

When running util.dumpInstance() with the 'ocimds' option set to true
users without password were not identified as potential issues, causing
the load operation to MDS instance to fail, as it has the
'validate_password' component enabled.

This issue has been fixed, the accounts which do not have passwords are
now listed as issues when the 'ocimds' option is used, while accounts
which are identified as roles (locked, with expired, empty password) are
dumped using the 'CREATE ROLE' statement. In order to fix these issues
and remove the accounts with an empty password from the dump, the
'skip_invalid_accounts' 'compatibility' option should be used.

BUG#32788788 WHEN FALLING BACK TO LOCK TABLES NOT ALL TABLES ARE LOCKED

When falling back to LOCK TABLES (i.e. because user does not have privileges
to run FLUSH TABLES WITH READ LOCK), multiple LOCK TABLES statements
were executed in the same session, which caused tables locked in the
previous LOCK TABLES statement to be unlocked.

The code has been refactored, session used by the main thread to gather
the information about schemas, tables, triggers, etc., is used to lock
the `mysql` system tables, while additional sessions are created to lock
the tables which are being dumped. The number of additional sessions
depends on the number of tables which need to be locked, and extra ones
will be created when LOCK TABLES statement exceeds the
max_allowed_packet limit.

Change-Id: I3f89486943ea34b7f6d274e290227e4fb57f429e