BUG#32734817 SHELL LOAD: LOAD DURING DUMP USING PAR GIVES UNAUTHORIZED ERROR
When a dump and load operation using PARs are running in parallel, the loader
gets authorization error trying to fetch the .idx file for a table. The problem
is caused because the handle to the .idx file is created using the table
file handle, which causes the loader to attempt downloading the .idx file to
using the wrong PAR.

The root cause is related to the fact that the IFile implementation for objects
in a manifest, have a wrong implementation of the parent() which should actually
return a handle to the manifest they belong.

This patch refactors the implementation of the manifest handling code with a
better separation of the operations related to the READ/WRITE modes and also
implement the correct behavior of the parent() function in the IFile implementation
for manifest objects.

Change-Id: I8e8ce43727a141e39b4f00d4bb4e813a26742ea4