BUG#32774826 ADMINAPI USES AND REFERS TO DEPRECATED REPLICATION SYSVARS AND OPTIONS
In 8.0.25, MySQL Replication replaced the terms "master", "slave", and
"mts" by "source", "replica", and "mta" in:

  - system variable names
  - status variables
  - SQL functions
  - command-line options (for server and all tools)
  - semisync plugin names
  - file names, C++ filenames and header guards, debug symbols, help
    texts
  - in performance_schema instrumentation names for locks,
    condition variables, threads, execution stages, allocators, and
    thread commands.

Since the AdminAPI uses and sets several replication sysvars, deprecation
warnings will be reported if the target instance version is >= 8.0.25.

This patch fixes that by ensuring the new terminology is used when
setting replication options to ensure no deprecation warnings are
issued. It also ensures deprecated sysvars are no longer set when no
need for it. Finally, some optimizations were added in the options
handling.

Updated existing tests to match the changes and added new test
scenarios.

Change-Id: I16c1f45157009b50b7bd6f4af37c0cf9993eee6b