BUG#32695301 DO NOT ACQUIRE LOCKS WHEN DRYRUN == TRUE
Previously, when running a dump with `dryRun` set to true, dumper would
attempt to lock the tables, interfering with the live database.

This patch changes the behaviour of the dumper, instead of trying to
acquire various locks and reacting to potential exceptions due to
missing privileges, it first checks if user has appropriate privilege to
acquire a lock and then continues with the execution depending on the
result of this check. This also allows to avoid acquiring the locks when
running in `dryRun` mode, while also reporting potential locking issues
(i.e. user lacks the privilege to execute FLUSH TABLES WITH READ LOCK,
and dumper will fall back to LOCK TABLES).

The User_privilege class used to obtain information about privileges has
been refactored: it now uses the SHOW GRANTS statement to fetch the
privileges user has (account + active roles). Support for partial
revokes has also been added.

Change-Id: Ife554b119fbd57f3859c5118855867f8f97a894a