Bug#32769580 ADMINAPI MAY UNNECESSARILY INVOKE UDF TO SET GR COMMUNICATION PROTOCOL
group_replication_set_communication_protocol() is a disruptive operation
that blocks the whole group while it's executing and is not recommended
to execute without explicit confirmation from the user.

But operations that change the cluster topology were performing upgrades
and downgrades automatically, which could cause problems on production
clusters (including a deadlock of the UDF).

The following changes were made:
- the automatic protocol upgrade was removed from
  Cluster.removeInstance, but a check was added with a note recommending
  an upgrade by the user using rescan()
- the protocol downgrade and check was rmovevd from Cluster.addInstance
  and rejoinInstance, although that code was never reached anyway,
  because it's impossible to join servers with older version.
- status() was augmented with a clusterErrors field which can be
  included when a protocol check indicates an upgrade is possible
- a new upgradeCommProtocol option was added to rescan(), which will
  upgrade the protocol of the group if possible.

Change-Id: I38f9e8e2a51ef3c630c26a52840e972bde4750ba