BUG#32593125 IMPROVE LOGGING FOR FAILED OCI OPERATIONS
When an OCI operation had failed, it was difficult to analyze the
problem, as by default not much information was written to the Shell
log.

In case of a failure, more information is now logged:
 - log_warning with the operation being executed (the target URL)
 - log_info with the request headers
 - log_info with the response headers (if available)
 - log_info with the response data (if available)

If the target URL is a PAR, the sensitive part of that URL is masked
when it is written to the log file.

Additionally, a bug in Shell has been fixed: when user was changing the
log level at runtime using i.e.: `shell.options.logLevel = 8`, the log
level value was changed, but the messages with the requested level were
not logged.

Change-Id: I2072c68025fffdef4c2a3c986fbe691701fb506c