WL#14577 Option to enable flushing of oversized trxs (sub-chunking)
A new option has been added to the util.loadDump():

maxBytesPerTransaction
----------------------

String, defaults to the bytesPerChunk value used during the dump.

Specifies the maximum number of bytes that can be loaded from a dump
data file per single LOAD DATA statement. If data file is bigger than
that, multiple LOAD DATA statements are executed per single file. If
this option is not specified explicitly (the default value is used),
only the files with data size greater than 1.5 * maxBytesPerTransaction
are affected.

This feature was introduced by a fix for BUG#31945539 (Sakila), with
the resume support implemented in a fix for BUG#31961688 (Sakila).

This WL allows users to overwrite the default, hard-coded value.

Change-Id: I74d17b348a1728b81c2853692f009c8af3260136