BUG#32926856 CAN'T LOAD DUMP WITH AUTO_INCREMENT VALUE OF 0
When loading a dump which contained a 0 value in AUTO_INCREMENT column,
data for that column was not loaded correctly.

If the 0 value is inserted into an AUTO_INCREMENT column, by default the
next sequence number is generated instead. When loading a dump, we want
to avoid this behaviour, as data may contain the literal 0, so we
disable the default behaviour using the 'NO_AUTO_VALUE_ON_ZERO' SQL
mode.

Change-Id: Ied21bf0f9ee5009603c491d265271b9922c97443