BUG#32934878 RACE CONDITION WHEN LOADING A DUMP WHICH IS IN PROGRESS
When loading a dump which is currently in progress, there can be a
situation where i.e. a table is marked as ready and selected for
further processing, while the metadata for that table has not been
downloaded/parsed yet. This can happen, because the flag, which
indicates that the file is ready, is set before the metadata is
actually read.

The code has been reordered to avoid this situation, all the relevant
flags are now volatile.

Change-Id: I4f7446037287aca880cba201d9e97ab57ddee940