BUG#32935004 SHELL: DUMP PRODUCES DUMP THAT CANNOT BE LOADED DUE TO DUPLICATED PRIMARY KEY
Dumps created by the current version of Shell were not loaded properly
by the older versions.

The fix for BUG#32850177 broke backwards compatibility of the table
metadata files: the 'primaryIndex' entry was moved to a different place
(and changed its type). The older versions of Shell were unable to find
the entry and assumed that table did not have the primary key, which
meant that the loader executed with 'createInvisiblePKs' = true was
trying to add another PK to a table which already had one.

Change-Id: I03866b37f2bc17ea2b59a3b25b2baaba9a7d85e7