BUG#32925914 DUMPS FROM MYSQL 5.6 MAY CONTAIN ROW_FORMAT=FIXED IN INNODB TABLES
When running a dump with the 'compatibility' option 'force_innodb'
enabled, the ENGINE option is changed to InnoDB, however the CREATE
TABLE statement was not checked for the option ROW_FORMAT=FIXED, which
is incompatible with an InnoDB table. This resulted in invalid CREATE
TABLE statements, and dumps failed to load.

This issue affected all server versions.

With this patch, the 'ocimds' option reports all tables which use
ROW_FORMAT=FIXED as errors, while the 'force_innodb' compatibility
option removes this clause from the CREATE TABLE statements.

Change-Id: I74893f6114dbb20cdabba6b4b6f3e92f20125317