WL#14632 Partition awareness in dump/load
This WL does not change any of the Dump&Load related APIs, it introduces
the partition awareness to the Dump&Load utilities, allowing for better
parallelization of these operations in case of partitioned tables.

This WL increases the major version number of the dump format, meaning
dumps created from this point forward cannot be loaded by the older
versions of Shell.

Change-Id: Id392a486cb3be10b25c861c59d0c0aedfb03366d