Fixed naming_style handling in multi-threaded code
The naming_style stack was being kept globally, which means different
threads could not use different languages at the same time. Sometimes,
race conditions would even cause issues when only Python is being used.

naming_style stack was changed to a thread_local variable and the
mutex for it was removed as it becomes unnecessary.

Change-Id: I4705b7e92b557a3e5396d75f0aaaf2291678ba1b