BUG#32987002 SHELL PACKAGE ON EL8 HAS INVALID DEPENDENCIES
We have found that defining runtime requirements for packages where Python
is not bundled is not required. Originally the dependencies were there
to ensure that the python runtime is available once the package is installed,
however we found that the build requirements will automatically add a
dependency to the python libs.

In distros using RPMs we identified that the libs contain the runtime
and the standard libraries while in distros using the DEBs the runtime
is also guaranteed because of an indirect dependency through the -stdlib.

Change-Id: Iee1602971f2bce0d8a2b867db16b503cd1e39a96