WL#14538 AdminAPI - ClusterSet Router management
This WL covers the process of re-assigning Routers to different
target Cluster from ClusterSets or change the routing policy regarding
invalidated clusters. It also covers the monitoring tasks of getting
the list of all Routers of a ClusterSet and the current options in
place.

List all Routers of a ClusterSet:
<ClusterSet>.listRouters([router])

List the Router options for all Routers or a specific Router of a
ClusterSet:
<ClusterSet>.routingOptions([router])

Change an option of all Routers of a ClusterSet or a specific Router
of a ClusterSet:
<ClusterSet>.setRoutingOption([router,] option, value)

Change-Id: I943152a2a10d4dfbf87b66b82683e078ceeaec5e