BUG#32969290 MYSQL SHELL : DUMNPINSTANCE() PREPERATION PHASE IS TAKING LOT OF TIME
When dumping an instance which has a lot of schemas and tables/views,
the "Gathering information" phase takes a lot of time.

The queries which gather information about table/view metadata were
using conditions to fetch only relevant data, which resulted in long
statements and poor response times from the server if there were a lot
of objects to filter. As a fix, filtering is limited to the input
provided by the user (i.e. options excludeSchemas, excludeTables, etc.).

Additional changes:

 - Added debug messages to trace execution time of each of the gathering
   information steps.
 - The dumper uploads the schema metadata files in threads.
 - Added per-thread cache of Oci_rest_service/Rest_service objects, which
   allows to use HTTP's keep-alive feature and reduces the overhead to
   establish a connection to the OCI service.
 - In some situations, HEAD request for an object stored in an OCI
   bucket was sent twice, unnecessary request has been removed.
 - Unnecessary calls to SHOW FIELDS when running a dump have been
   removed. As a consequence, the EXECUTE privilege is no longer
   required when dumping a view which uses a function to get data
   (applies to 5.7, see BUG#32865281 for more info.)

Note: "Gathering information" step is a step where schemas, tables,
      views to be dumped are selected and their metadata required by the
      dumper is fetched from the server.

Change-Id: Ib2ebe442fe840dd9a93798b972e572570cbaaa85