BUG#31645798 BUG#32561035 Fixes for util.loadDump()
BUG#31645798 SHELL DUMP & LOAD: WHEN LOADING 16K TABLES, PROGRESS IS NOT SHOWN

When a dump with thousands of files was loaded from the OCI bucket, user
might have an impression that the operation was stuck, as no progress
was reported when the metadata files were scanned.

This patch adds the progress reporting to these steps, as well as
improves the performance of loading the metadata/DDL files.

BUG#32561035 LOAD SHOULD ENSURE PROGRESSFILE IS VALID FOR THE TARGET SERVER

The progress file now includes the UUID of the target server. If the
load operation is resumed, and the saved UUID is different from the
current UUID, an error is reported and the load operation is aborted.

-----------------------------------------------------------------------

Performance improvements for util.loadDump():
 - The loader's rescan for metadata files operation is now
   multithreaded.
 - The rescan metadata step and DDL load step are now using an
   additional pool of threads to fetch the contents of files. The size
   of that pool is controlled be a new option `backgroundThreads`, by
   default it is not set. In such case loader will use the value of
   `threads` option (local dump), or four times that value (non-local
   dump).

Additional changes:
 - The loader will now store the timestamp (in milliseconds) of each
   operation saved in the progress log.

Change-Id: Ib7e47f34fd35439d8519c57b6ab68a9d4936fddc