BUG#33079172 BUG#33087120 fixes for 'strip_definers' compatibility option
BUG#33079172 SQL SECURITY INVOKER WILL BE INSERTED ON WRONG LOCATION USING UTIL.DUMPSCHEMAS

When dumping a procedure or function which contained only a simple
statement (in the reported case: IF) and using the 'strip_definers'
compatibility option, the dumped CREATE statement had "SQL SECURITY
INVOKER" clause inserted at wrong location, making it impossible to
load.

The handling of the simple statements has been fixed, all of them are
now properly recognized.

BUG#33087120 ENSURE THAT `SQL SECURITY INVOKER` CLAUSE IS ALWAYS ADDED

The implementation of 'strip_definers' compatibility option in some
circumstances would not add the `SQL SECURITY INVOKER` clause, this is
now fixed. If a function/procedure/view is missing an `SQL SECURITY
INVOKER` clause and `ocimds` option is used, this fact is reported as a
separate issue (previously was not reported). If the 'strip_definers'
compatibility option is used to fix this problem, user is notified by a
separate entry (previously this notification was a part of a message
which talked about the DEFINER clause).

Change-Id: Ic43391613293302b7bf104d6ba989844066312ce