BUG#33096667 HANDLING OF CTRL+C ON THE SHELL IS INCONSISTENT
When shell was displaying a prompt/password prompt, sending a SIGINT
signal to the Shell's process did not interrupt the prompt.

The non-Windows handling of SIGINT signal was inconsistent, some code
was using signal() while other was using sigaction(). Code has been
adjusted to use sigaction() everywhere, the SA_RESTART flag has been
removed, so that it is possible to interrupt system calls instead of
restarting them. This change restores behaviour introduced by the fix
for BUG#27894642.

Change-Id: I24e503fd755e78a42fadae1ab7679ed17024ba20