BUG#33144419 MYSQLSHELL UTIL.LOADDUMP() DOESN'T RECOGNISE USERS LIKE 'USERNAME'@'1.2.3.4'
Loading a dump using an account which had the host part specified using
an IPv4 address and a netmask failed with an invalid syntax error.

When the load operation is initialized, the account of the user which is
executing that operation is excluded from the list of loaded accounts.
This step was not handling the netmask properly, which resulted in the
error described above.

Change-Id: Ibb3a70aca4ac5504d64bb35c8126a521892afb96