BUG#32928528 MYSQL SHELL LOADDUMP FAILS WITH DEADLOCK WHEN RECREATING TABLES REFERENCED BY VIEWS.
When executing DDL for tables which are referenced by existing views,
the load operation could fail with a deadlock error.

In case of deadlock errors, the statement was re-executed up to 20
times, with a constant delay between attempts equal to 200 ms. On a slow
server, this could be not enough time to avoid the deadlocks. The retry
mechanism has been rewritten, the delay is now doubled after each
attempt, while the number of attempts is bounded by the total waiting
time which is set to 5 minutes.

Change-Id: I52d43b1681ef5249455291db21ccf9d3fca57e34