BUG#32955616 TABLES WITH AVERAGE ROW SIZE LARGER THAN BYTESPERCHUNK DON'T GET CHUNKED AS EXPE
When table has rows which are bigger than `bytesPerChunk`, it is
expected that each such row will end up in a separate chunk file.

The chunker algorithm was using wrong assumptions regarding the minimum
step size, which caused that at least two rows were written to the chunk
file. Additionally, when the value of the index column was close to the
maximum, the algorithm was finished and the last chunk contained even
more rows. These problems are now fixed, the expected behaviour is now
ensured.

Change-Id: I6ed70f47bb5673e8df466e5e28f4877b4a85e7a7