BUG#33122234 BUG#33155373 Fixes for util.loadDump() with PARs
BUG#33122234 CAN'T LOAD DUMP WITH >1K FILES FROM PREFIX PAR

Loading a dump with more than 1000 files from a prefix PAR failed or
reported that nothing was loaded.

Listing a directory from a prefix PAR did not support pagination and
returned only up to 1000 first files, this is now fixed.

BUG#33155373 mysqlsh reports Directory handling for https protocol

When loading a dump using a PAR, if PAR URL contained a typo, it was
not recognized by the loader as a PAR and resulted in a cryptic error
message. The error message was updated, in case of an invalid PAR the
expected URL format is printed out.

Additional fixes:
 - PAR URLs are always anonymized when printed to either console or log,
   the PAR ID is replaced with the word '<secret>'.
 - When using a prefix PAR, the Rest_service objects are reused if the
   same base URL is requested from the same thread, allowing to utilize
   the Keep-Alive feature of HTTP protocol.

Change-Id: I03ce4a86af0c62f6a9443c224131b80e8e8e2a44