BUG#32799076 SHELL UTIL.LOADDUMP() EXCLUDES ALL USERS INSTEAD OF SPECIFIED ONLY
The documentation on the excludeUsers/includeUsers options of the
util.dumpInstance() and util.loadDump() utilities was wrong, it stated
that when a user is specified as `user_name`, 'user_name'@'%' is
assumed.

The documentation has been fixed, when `user_name` is given, all
accounts with the specified user name are included/excluded.

Additional fix:
 - When loading a dump, certain accounts should always be excluded,
   while shell mistakingly excluded them only if the excludeUsers option
   was explicitly given by the user. This is now fixed.

Change-Id: I5c93ec76cf11cc0895b921641992aaf336cdc4a7